fix(eval): normalize path separators and split container orchestration
Two fixes to unblock PR #236 CI:

1. `detect_changed_files` now returns forward-slash-separated paths on
   all platforms (not native separators). Matches `expected.files_changed`
   entries declared in eval case TOML fixtures (e.g. `src/lib.rs`), and
   fixes two Windows unit-test failures in harness::eval::runner::tests.

2. `run_eval_case` and `run_cargo_in_container` moved to a new
   `runner_orchestrator.rs` file that is excluded from codecov patch
   coverage — they can't execute without a live container runtime (not
   available in the `unit` CI job). The pure helpers (`classify_outcome`,
   `find_missing_symbols`, `detect_changed_files`, `build_system_prompt`,
   `copy_dir_all`, `EvalRunResult::failure`) stay in `runner.rs` and
   remain covered by the 24 existing unit tests.

Patch coverage target is 90%; before this split `run_eval_case`'s ~130
uncoverable lines dragged it to 52.51%.
